@@ -14,17 +14,14 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.1",
-    "@docusaurus/preset-classic": "^2.0.1",
-    "@mdx-js/react": "^1.6.22",
-    "@svgr/webpack": "^5.5.0",
-    "classnames": "^2.2.6",
-    "clsx": "^1.2.1",
-    "file-loader": "^6.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-table": "^7.8.0",
-    "url-loader": "^4.1.1"
+    "@docusaurus/core": "^3.1.1",
+    "@docusaurus/preset-classic": "^3.1.1",
+    "@mdx-js/react": "^3.0.0",
+    "@tanstack/react-table": "^8.13.2",
+    "clsx": "^2.0.0",
+    "prism-react-renderer": "^2.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "browserslist": {
     "production": [
