@@ -1,7 +1,7 @@
 /**
  * Import this file in your Docusaurus `sidebars.js` file.
  *
- * Auto-generated by Alt3.Docusaurus.PowerShell.
+ * Auto-generated by Alt3.Docusaurus.Powershell 1.0.34.
  *
  * Copyright (c) 2019-present, ALT3 B.V.
  *
